[Fix] Resolve errors - Changed anime migration relations to unsignedBigInteger - Updated composer

+++ database/migrations/2018_08_17_103205_create_anime_studio_table.php
@@ -16,7 +16,7 @@ class CreateAnimeStudioTable extends Migration
     {
         Schema::create(AnimeStudio::TABLE_NAME, function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('anime_id');
+            $table->unsignedBigInteger('anime_id');
             $table->unsignedBigInteger('studio_id');
             $table->timestamps();
         });
